verify notify wechar

+++ src/main/java/com/top1/marketinfo/entity/Demand.java
@@ -37,7 +37,7 @@ public class Demand {
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private Date invalidDate = new Date();
 
-    private int publisherId;
+    private long publisherId;
 
     private String publisherNickName;
